requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.5
djangorestframework==3.13.1
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
